de-duplicate code from tools.js
